requirements.txt
Flask==3.0.0
PyYAML==6.0.1
requests==2.31.0
regex==2023.12.25
